refactor: extract logout logic to custom hook
